Move current MinecraftInterface to legacy package; add base for new interface
