+++ src/main/java/amidst/mojangapi/minecraftinterface/MinecraftInterfaces.java
@@ -15,6 +15,8 @@ import amidst.logging.AmidstLogger;
 import amidst.mojangapi.file.LauncherProfile;
 import amidst.mojangapi.minecraftinterface.legacy.LegacyClassTranslator;
 import amidst.mojangapi.minecraftinterface.legacy.LegacyMinecraftInterface;
+import amidst.mojangapi.minecraftinterface.legacy._1_13ClassTranslator;
+import amidst.mojangapi.minecraftinterface.legacy._1_13MinecraftInterface;
 import amidst.mojangapi.minecraftinterface.local.DefaultClassTranslator;
 import amidst.mojangapi.minecraftinterface.local.LocalMinecraftInterface;
 
@@ -44,9 +46,16 @@ public enum MinecraftInterfaces {
     }
 
     @NotNull
+    public static ClassTranslator getClassTranslatorFromVersion(RecognisedVersion recognisedVersion) {
+        return fromVersion(recognisedVersion).classTranslator;
+    }
+
+    @NotNull
     private static Factory fromVersion(RecognisedVersion version) {
         if(RecognisedVersion.isOlderOrEqualTo(version, LegacyMinecraftInterface.LAST_COMPATIBLE_VERSION)) {
             return new Factory(LegacyClassTranslator.get(), LegacyMinecraftInterface::new);
+        } else if(RecognisedVersion.isOlderOrEqualTo(version, _1_13MinecraftInterface.LAST_COMPATIBLE_VERSION)) {
+            return new Factory(_1_13ClassTranslator.get(), _1_13MinecraftInterface::new);
         } else {
             return new Factory(DefaultClassTranslator.get(), LocalMinecraftInterface::new);
         }
