+++ src/main/java/amidst/mojangapi/minecraftinterface/legacy/_1_13SymbolicNames.java
@@ -1,0 +1,43 @@
+package amidst.mojangapi.minecraftinterface.legacy;
+
+import amidst.documentation.Immutable;
+import amidst.mojangapi.minecraftinterface.local.SymbolicNames;
+
+@Immutable
+public enum _1_13SymbolicNames {
+	;
+
+    public static final String CLASS_WORLD_TYPE = SymbolicNames.CLASS_WORLD_TYPE;
+    public static final String FIELD_WORLD_TYPE_DEFAULT = SymbolicNames.FIELD_WORLD_TYPE_DEFAULT;
+    public static final String FIELD_WORLD_TYPE_FLAT = SymbolicNames.FIELD_WORLD_TYPE_FLAT;
+    public static final String FIELD_WORLD_TYPE_LARGE_BIOMES = SymbolicNames.FIELD_WORLD_TYPE_LARGE_BIOMES;
+    public static final String FIELD_WORLD_TYPE_AMPLIFIED = SymbolicNames.FIELD_WORLD_TYPE_AMPLIFIED;
+    public static final String FIELD_WORLD_TYPE_CUSTOMIZED = SymbolicNames.FIELD_WORLD_TYPE_CUSTOMIZED;
+
+	public static final String CLASS_BOOTSTRAP = "Bootstrap";
+	public static final String METHOD_BOOTSTRAP_REGISTER = "register";
+	public static final String METHOD_BOOTSTRAP_REGISTER2 = "register2";
+	public static final String METHOD_BOOTSTRAP_REGISTER3 = "register3";
+
+	public static final String CLASS_LAYER_UTIL = "LayerUtil";
+	public static final String METHOD_LAYER_UTIL_INITIALIZE_ALL = "initializeAll";
+
+	public static final String CLASS_GEN_SETTINGS = "OverworldGenSettings";
+	public static final String CONSTRUCTOR_GEN_SETTINGS = "<init>";
+
+	public static final String CLASS_GEN_LAYER = "GenLayer";
+	public static final String METHOD_GEN_LAYER_GET_BIOME_DATA = "getBiomeData";
+	public static final String METHOD_GEN_LAYER_GET_BIOME_DATA2 = "getBiomeData2";
+
+	public static final String CLASS_BIOME = "Biome";
+	public static final String METHOD_BIOME_GET_ID = "getBiomeId";
+
+	public static final String CLASS_REGISTRY = "Registry";
+	public static final String FIELD_REGISTRY_META_REGISTRY = "metaRegistry";
+	public static final String METHOD_REGISTRY_GET_BY_KEY = "getByKey";
+	public static final String METHOD_REGISTRY_GET_BY_KEY2 = "getByKey2";
+	public static final String METHOD_REGISTRY_GET_ID = "getId";
+
+	public static final String CLASS_REGISTRY_KEY = "RegistryKey";
+	public static final String CONSTRUCTOR_REGISTRY_KEY = "<init>";
+}
